后端最后功能完成

+++ src/main/java/com/aks/cateringinfosys/service/impl/CommentServiceImpl.java
@@ -3,11 +3,15 @@ package com.aks.cateringinfosys.service.impl;
 import cn.hutool.json.JSONUtil;
 import com.aks.cateringinfosys.dto.Result;
 import com.aks.cateringinfosys.entry.Comment;
+import com.aks.cateringinfosys.entry.User;
 import com.aks.cateringinfosys.mappers.CommentMapper;
 import com.aks.cateringinfosys.mappers.ImageMapper;
+import com.aks.cateringinfosys.mappers.UserMapper;
 import com.aks.cateringinfosys.service.ICommentService;
 import com.aks.cateringinfosys.utils.RedisIdWorker;
 import com.aks.cateringinfosys.utils.UserHolder;
+import com.github.pagehelper.PageHelper;
+import com.github.pagehelper.PageInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -41,24 +45,34 @@ public class CommentServiceImpl implements ICommentService {
     CommentMapper commentMapper;
     @Autowired
     ImageMapper imageMapper;
+    @Autowired
+    UserMapper userMapper;
     @Resource
     StringRedisTemplate stringRedisTemplate;
 
 
+    /**
+     * 根据店铺id获取评论
+     * @param rid
+     * @return
+     */
     @Override
-    public Result getRestComments(Integer rid) {
+    public Result getRestComments(Long rid) {
+        // todo 在redis中获取指定店铺的评论列表
         String commentStr = stringRedisTemplate.opsForValue().get(CACHE_COMMENTLIST_KEY + rid);
-        if (commentStr != null && commentStr != null && !CACHE_NULL.equals(commentStr)) {
-            List<Comment> commentList = JSONUtil.toBean(commentStr, List.class);
+        // todo 判断评论为空字符串或者为设置的 is null
+        if (commentStr != null && !CACHE_NULL.equals(commentStr)) { // todo 非空则输出
+            List<Comment> commentList = JSONUtil.toList(commentStr, Comment.class);
             logger.info(UserHolder.getUser().getUid() + "获取店铺"+ rid + "评论"+commentList);
             return Result.ok(commentList);
         }
-        if (commentStr.equals(CACHE_NULL)) {
+        // todo 为空返回前端为空
+        if (CACHE_NULL.equals(commentStr)) {
             logger.info(UserHolder.getUser().getUid() + "获取店铺"+ rid + "评论为空");
             return Result.fail("用户评论为空");
         }
         List<Comment> commentList = commentMapper.queryCommentListByRestId(rid);
-        if(commentList == null) {
+        if(commentList == null || commentList.size() == 0) {
             stringRedisTemplate.opsForValue().set(CACHE_COMMENTLIST_KEY +rid,
                     CACHE_NULL,
                     CACHE_NULL_TTL,
@@ -66,22 +80,32 @@ public class CommentServiceImpl implements ICommentService {
             logger.info(UserHolder.getUser().getUid() + "获取店铺"+ rid + "评论为空");
             return Result.fail("用户评论为空");
         }
-        commentList.
+        // todo 流操作
+        List<Object> collect = commentList.
                 stream().
                 map(comment -> {
+                    String nickName = userMapper.queryNameByUid(comment.getUid());
                     List<String> imageList = imageMapper.queryImageListByForeign(comment.getCid());
                     comment.setImageList(imageList);
+                    comment.setNickName(nickName);
                     return null;
                 }).collect(Collectors.toList());
         stringRedisTemplate.opsForValue().set(CACHE_COMMENTLIST_KEY+rid,
-                JSONUtil.toJsonStr(commentList),
+                JSONUtil.toJsonStr(collect),
                 CACHE_COMMENTLIST_TTL,
                 TimeUnit.MINUTES);
 
         logger.info(UserHolder.getUser().getUid() + "获取店铺"+ rid + "评论"+commentList);
-        return Result.ok(commentList);
+        return Result.ok(collect);
     }
 
+    /**
+     * 对指定评论点赞，由于点赞的数量对于数据的要求并不严格，并不需要进行严格的数据一致性要求，
+     * 无需删除redis中的数据，采用过期后自动更新
+     * @param cid
+     * @return
+     */
+    @Transactional
     @Override
     public Result likeComment(Integer cid) {
 
@@ -109,12 +133,16 @@ public class CommentServiceImpl implements ICommentService {
         }
     }
 
+    /**
+     * 为指定店铺添加评论
+     * @param comment
+     * @return
+     */
     @Transactional
     @Override
     public Result addComment(Comment comment) {
         Long useId = UserHolder.getUser().getUid();
         comment.setUid(useId);
-        comment.setLike(0);
         comment.setCreateTime(LocalDateTime.now());
         long cid = idWorker.nextId();
         comment.setCid(cid);
@@ -141,6 +169,10 @@ public class CommentServiceImpl implements ICommentService {
         return Result.ok("添加评论成功");
     }
 
+    /**
+     * 获取自己的所有评论
+     * @return
+     */
     @Override
     public Result getSelfComment() {
         Long id = UserHolder.getUser().getUid();
@@ -150,8 +182,8 @@ public class CommentServiceImpl implements ICommentService {
             logger.info("用户"+id+"获取个人评论" + s);
             return Result.ok(commentList);
         }
-        if (s.equals(CACHE_NULL)) {
-            logger.info("用户"+id+"获取个人评论失败");
+        if (CACHE_NULL.equals(s)) {
+            logger.info("用户"+id+"获取个人评论为空");
             return Result.fail("你尚未评论,去评论试试吧");
         }
         List<Comment> commentList = commentMapper.queryCommentListByUserId(id);
@@ -163,7 +195,8 @@ public class CommentServiceImpl implements ICommentService {
             logger.info("用户" + id + "获取评论为空");
             return Result.fail("你尚未评论,去评论试试吧");
         }
-        commentList.
+        //流操作获取所有的评论照片
+        List<Object> collect = commentList.
                 stream().
                 map(comment -> {
                     List<String> imageList = imageMapper.queryImageListByForeign(comment.getCid());
@@ -171,25 +204,81 @@ public class CommentServiceImpl implements ICommentService {
                     return null;
                 }).collect(Collectors.toList());
         stringRedisTemplate.opsForValue().set(CACHE_COMMENTLIST_KEY+id,
-                JSONUtil.toJsonStr(commentList),
+                JSONUtil.toJsonStr(collect),
                 CACHE_COMMENTLIST_TTL,
                 TimeUnit.MINUTES);
 
-        logger.info("用户"+id+"获取个人评论" + commentList);
-        return Result.ok(commentList);
+        logger.info("用户"+id+"获取个人评论" + collect);
+        return Result.ok(collect);
     }
 
+    /**
+     * 删除指定id的评论
+     * @param cid
+     * @return
+     */
     @Override
     @Transactional
     public Result deleteCommentById(Long cid) {
+        Comment comment = commentMapper.queryCommentByCid(cid);
+        if (comment == null ) {
+            return Result.fail("评论已经不存在啦");
+        }
         Integer flag = commentMapper.deleteCommentById(cid);
         if (flag != 1) {
             logger.info("用户" + UserHolder.getUser().getUid()+ "删除一条评论" + cid+"失败");
             throw new RuntimeException("删除评论失败，请重试");
         }
         logger.info("用户" + UserHolder.getUser().getUid()+ "删除一条评论" + cid);
+        stringRedisTemplate.delete(CACHE_COMMENTLIST_KEY+comment.getUid());
+        stringRedisTemplate.delete(CACHE_COMMENTLIST_KEY+comment.getRid());
         stringRedisTemplate.delete(CACHE_COMMENTLIST_KEY + UserHolder.getUser().getUid());
         return Result.ok("删除评论成功");
+    }
+
+    @Override
+    public Result getCommentList(Integer type, Long id, Integer currentPage, Integer pageSize) {
+        String key = String.valueOf(currentPage << 10 + pageSize);
+        String commentStr = (String) stringRedisTemplate.opsForHash().get(CACHE_COMMENTLIST_KEY + id, key);
+        if (commentStr != null && !CACHE_NULL.equals(commentStr)) {
+            List<Comment> comments = JSONUtil.toList(commentStr, Comment.class);
+            return Result.ok(comments);
+        }
+        if (CACHE_NULL.equals(commentStr)) {
+            return Result.fail("查询的信息为空");
+        }
+        List<Comment> comments = null;
+
+        PageHelper.startPage(currentPage,pageSize);
+        PageInfo<Comment> pageInfo = null;
+        if (type == 1) {
+            comments = commentMapper.queryCommentListByUserId(id);
+            pageInfo = new PageInfo<>(comments);
+        } else if (type == 2) {
+            comments = commentMapper.queryCommentListByRestId(id);
+            pageInfo = new PageInfo<>(comments);
+        } else {
+            comments = commentMapper.queryCommentListAll();
+            pageInfo = new PageInfo<>(comments);
+        }
+        comments = pageInfo.getList();
+        if (comments == null || comments.size() == 0) {
+            stringRedisTemplate.opsForHash().put(CACHE_COMMENTLIST_KEY+id,key,CACHE_NULL);
+            stringRedisTemplate.expire(CACHE_COMMENTLIST_KEY+id,CACHE_NULL_TTL,TimeUnit.MINUTES);
+            return Result.fail("查询的评论为空");
+        }
+        comments.stream().map(comment -> {
+            Long uid = comment.getUid();
+            User user = userMapper.queryUserByUserId(uid);
+            if (user != null) {
+                comment.setNickName(user.getNickName());
+            }
+            return null;
+        }).collect(Collectors.toList());
+        stringRedisTemplate.opsForHash().put(CACHE_COMMENTLIST_KEY+id,key,JSONUtil.toJsonStr(comments));
+        stringRedisTemplate.expire(CACHE_COMMENTLIST_KEY+id,CACHE_COMMENTLIST_TTL,TimeUnit.MINUTES);
+
+        return Result.ok(comments);
     }
 
 }
